Move some useful JS call graph code to com.ibm.wala.cast.js
