+++ com.ibm.wala.cast.js.test/harness-src/com/ibm/wala/cast/js/test/TestAjaxsltCallGraphShape.java
@@ -10,7 +10,8 @@
  */
 package com.ibm.wala.cast.js.test;
 
-import com.ibm.wala.cast.js.test.JSCallGraphBuilderUtil.CGBuilderType;
+import com.ibm.wala.cast.js.util.JSCallGraphBuilderUtil;
+import com.ibm.wala.cast.js.util.JSCallGraphBuilderUtil.CGBuilderType;
 import com.ibm.wala.ipa.callgraph.CallGraph;
 import com.ibm.wala.util.CancelException;
 import com.ibm.wala.util.WalaException;
