+++ com.ibm.wala.cast.js/source/com/ibm/wala/cast/js/util/FieldBasedCGUtil.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  */
-package com.ibm.wala.cast.js.test;
+package com.ibm.wala.cast.js.util;
 
 import com.ibm.wala.cast.ipa.callgraph.CAstAnalysisScope;
 import com.ibm.wala.cast.ir.ssa.AstIRFactory;
